format, separated get_current_user from routes, changed location of error message renderer
